refactor: 타이머 로직을 useBattleTimer 커스텀 훅으로 분리
